Fix launch not working in some cases with spaces in program path
closes #17

I couldn't find a way to to escape paths passed to fdb's `run` command (neither quoting the path nor `^ ` work).

The workaround is to execute the `run` command without arguments and launch the swf manually, which will then connect to fdb. This also seems to be what FlashDevelop is doing.